+++ cosmate-backend/src/main/java/com/cosmate/service/impl/CostumeServiceImpl.java
@@ -45,7 +45,7 @@ public class CostumeServiceImpl implements CostumeService {
 
     @Override
     @Transactional
-    public CostumeResponse updateCostume(Long id, CostumeRequest request) {
+    public CostumeResponse updateCostume(Integer id, CostumeRequest request) {
         Costume costume = costumeRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy bộ đồ ID: " + id));
 
@@ -167,10 +167,10 @@ public class CostumeServiceImpl implements CostumeService {
     }
 
     @Override public List<CostumeResponse> getAllCostumes() { return costumeRepository.findAll().stream().map(this::mapToResponse).collect(Collectors.toList()); }
-    @Override public CostumeResponse getById(Long id) { return mapToResponse(costumeRepository.findById(id).orElseThrow()); }
+    @Override public CostumeResponse getById(Integer id) { return mapToResponse(costumeRepository.findById(id).orElseThrow()); }
     @Override
     @Transactional
-    public void deleteCostume(Long id) {
+    public void deleteCostume(Integer id) {
         Costume costume = costumeRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy bộ đồ để xóa!"));
